fix service package hierarchy
