+++ internal/service/account/account_test.go
@@ -1,4 +1,4 @@
-package service
+package account
 
 import (
 	"context"
