+++ internal/service/order/order_test.go
@@ -1,4 +1,4 @@
-package service
+package order
 
 import (
 	"context"
